fix: resolve 11 GitHub issues (#368-#378)
CRITICAL:
- #368: payments.service.ts already used Prisma.Decimal (confirmed OK)
- #369: add Serializable isolation to autoAllocatePayment + applyCreditBalance
- #370: add IMEI duplicate check inside contract creation transaction

HIGH:
- #371: add composite indexes to Customer, Payment, Contract models
- #372: remove --\s*$ XSS regex (false positive on Thai notes)
- #373: extract status labels to constants/statusLabels.ts, remove duplication
- #374: split NotificationsPage.tsx (963 lines) into 5 sub-components

MEDIUM:
- #375: register GlobalHttpExceptionFilter in main.ts
- #376: skipped — requires business requirement confirmation (flat rate is correct)
- #377: add token expiry warning toast 2 min before session expires
- #378: add NotFoundPage, replace catch-all redirect with proper 404 page
