+++ apps/web/src/pages/ContractsPage.tsx
@@ -11,6 +11,7 @@ import { Card, CardContent } from '@/components/ui/card';
 import { toast } from 'sonner';
 
 import { exportToExcel } from '@/utils/excel.util';
+import { CONTRACT_STATUS_LABELS as statusLabels } from '@/constants/statusLabels';
 
 interface Contract {
   id: string;
@@ -32,16 +33,6 @@ interface Contract {
   _count: { payments: number; contractDocuments: number };
 }
 
-const statusLabels: Record<string, { label: string; className: string }> = {
-  DRAFT: { label: 'ร่าง', className: 'bg-muted text-foreground' },
-  ACTIVE: { label: 'ผ่อนอยู่', className: 'bg-green-100 text-green-700' },
-  OVERDUE: { label: 'ค้างชำระ', className: 'bg-yellow-100 text-yellow-700' },
-  DEFAULT: { label: 'ผิดนัด', className: 'bg-red-100 text-red-700' },
-  EARLY_PAYOFF: { label: 'ปิดก่อน', className: 'bg-primary-100 text-primary-700' },
-  COMPLETED: { label: 'ครบ', className: 'bg-teal-100 text-teal-700' },
-  EXCHANGED: { label: 'เปลี่ยนเครื่อง', className: 'bg-primary-100 text-primary-700' },
-  CLOSED_BAD_DEBT: { label: 'หนี้สูญ', className: 'bg-red-200 text-red-800' },
-};
 
 interface PaginatedResponse<T> {
   data: T[];
